Model moved to new widget
