+++ js/ui/file_manager/ui.file_manager.items_list.thumbnails.list_box.js
@@ -1,23 +1,48 @@
+import $ from '../../core/renderer';
 import { extend } from '../../core/utils/extend';
+import messageLocalization from '../../localization/message';
+import { getDisplayFileSize } from './ui.file_manager.common';
 import { BindableTemplate } from '../../core/templates/bindable_template';
 
 import CollectionWidget from '../collection/ui.collection_widget.edit';
 
 const FILE_MANAGER_THUMBNAILS_ITEM_CLASS = 'dx-filemanager-thumbnails-item';
+const FILE_MANAGER_THUMBNAILS_ITEM_NAME_CLASS = 'dx-filemanager-thumbnails-item-name';
+const FILE_MANAGER_THUMBNAILS_ITEM_SPACER_CLASS = 'dx-filemanager-thumbnails-item-spacer';
 
 const FILE_MANAGER_THUMBNAILS_ITEM_CLASS_DATA_KEY = 'dxFileManagerItemDataKey';
 
 class FileManagerThumbnailListBox extends CollectionWidget {
+    _init() {
+        super._init();
+        this._initActions();
+        this._$scrollable = this.option('scrollableElement');
+    }
+
+    _initActions() {
+        this._onItemEnterKeyPressed = this._createActionByOption('onItemEnterKeyPressed');
+    }
+
+    _initTemplates() {
+        super._initTemplates();
+        this._getItemThumbnailContainer = this.option('getItemThumbnail');
+        this._templateManager.addDefaultTemplates({
+            item: new BindableTemplate(function($container, data, itemModel) {
+                const $itemElement = this._getDefaultItemTemplate(itemModel, $container);
+                $container.append($itemElement);
+            }.bind(this), ['fileItem'], this.option('integrationOptions.watchMethod'))
+        });
+    }
 
     _supportedKeys() {
         return extend(super._supportedKeys(), {
             upArrow(e) {
                 this._beforeKeyProcessing(e);
-                this._processMoveArrow(-1, false, e);
+                this._processArrowKeys(-1, false, e);
             },
             downArrow(e) {
                 this._beforeKeyProcessing(e);
-                this._processMoveArrow(1, false, e);
+                this._processArrowKeys(1, false, e);
             },
             home(e) {
                 this._beforeKeyProcessing(e);
@@ -25,7 +50,7 @@ class FileManagerThumbnailListBox extends CollectionWidget {
             },
             end(e) {
                 this._beforeKeyProcessing(e);
-                this._processHomeEndKeys(this.getItems().length - 1, true, e);
+                this._processHomeEndKeys(this._getItemsLength() - 1, true, e);
             },
             pageUp(e) {
                 this._beforeKeyProcessing(e);
@@ -48,13 +73,57 @@ class FileManagerThumbnailListBox extends CollectionWidget {
         });
     }
 
-    _initActions() {
-        this._getDefaultItemTemplate = this.option('defaultItemTemplate');
-        this._onItemEnterKeyPressed = this._createActionByOption('onItemEnterKeyPressed');
-        this._beforeKeyProcessing = this.option('beforeKeyProcessing');
-        this._processMoveArrow = this.option('processMoveArrow');
-        this._processHomeEndKeys = this.option('processHomeEndKeys');
-        this._processPageChange = this.option('processPageChange');
+    _beforeKeyProcessing(e) {
+        e.preventDefault();
+        this._resetLayoutModel();
+    }
+
+    _processArrowKeys(offset, horizontal, eventArgs) {
+        const item = this._getFocusedItem();
+        if(item) {
+            if(!horizontal) {
+                const layout = this._getLayoutModel();
+                if(!layout) {
+                    return;
+                }
+
+                offset *= layout.itemPerRowCount;
+            }
+
+            const newItemIndex = this._getIndexByItem(item) + offset;
+            this._focusItemByIndex(newItemIndex, true, eventArgs);
+        }
+    }
+
+    _processHomeEndKeys(index, scrollToItem, eventArgs) {
+        this._focusItemByIndex(index, scrollToItem, eventArgs);
+    }
+
+    _processPageChange(pageUp, eventArgs) {
+        const item = this._getFocusedItem();
+        if(!item) {
+            return;
+        }
+
+        const layout = this._getLayoutModel();
+        if(!layout) {
+            return;
+        }
+
+        const itemLayout = this._createItemLayoutModel(this._getIndexByItem(item));
+
+        const rowOffset = pageUp ? layout.rowPerPageRate : -layout.rowPerPageRate;
+        const newRowRate = itemLayout.itemRowIndex - rowOffset;
+        const roundFunc = pageUp ? Math.ceil : Math.floor;
+        const newRowIndex = roundFunc(newRowRate);
+        let newItemIndex = newRowIndex * layout.itemPerRowCount + itemLayout.itemColumnIndex;
+        if(newItemIndex < 0) {
+            newItemIndex = 0;
+        } else if(newItemIndex >= this._getItemsLength()) {
+            newItemIndex = this._getItemsLength() - 1;
+        }
+
+        this._focusItemByIndex(newItemIndex, true, eventArgs);
     }
 
     _itemContainer() {
@@ -69,22 +138,36 @@ class FileManagerThumbnailListBox extends CollectionWidget {
         return FILE_MANAGER_THUMBNAILS_ITEM_CLASS_DATA_KEY;
     }
 
-    _initTemplates() {
-        super._initTemplates();
-        this._initActions();
-        this._templateManager.addDefaultTemplates({
-            item: new BindableTemplate(function($container, data, itemModel) {
-                const $itemElement = this._getDefaultItemTemplate(itemModel, $container);
-                $container.append($itemElement);
-            }.bind(this), ['fileItem'], this.option('integrationOptions.watchMethod'))
-        });
+    _getDefaultItemTemplate(fileItemInfo, $itemElement) {
+        $itemElement.attr('title', this._getTooltipText(fileItemInfo));
+
+        const $itemThumbnail = this._getItemThumbnailContainer(fileItemInfo);
+
+        const $itemSpacer = $('<div>').addClass(FILE_MANAGER_THUMBNAILS_ITEM_SPACER_CLASS);
+
+        const $itemName = $('<div>')
+            .addClass(FILE_MANAGER_THUMBNAILS_ITEM_NAME_CLASS)
+            .text(fileItemInfo.fileItem.name);
+
+        $itemElement.append($itemThumbnail, $itemSpacer, $itemName);
+    }
+
+    _getTooltipText(fileItemInfo) {
+        const item = fileItemInfo.fileItem;
+        if(item.tooltipText) {
+            return item.tooltipText;
+        }
+
+        let text = `${item.name}\r\n`;
+        if(!item.isDirectory) {
+            text += `${messageLocalization.format('dxFileManager-listThumbnailsTooltipTextSize')}: ${getDisplayFileSize(item.size)}\r\n`;
+        }
+        text += `${messageLocalization.format('dxFileManager-listThumbnailsTooltipTextDateModified')}: ${item.dateModified}`;
+        return text;
     }
 
     _itemSelectHandler(e) {
         const index = this.getIndexByItemElement(e.currentTarget);
-        if(index === -1) {
-            return;
-        }
         let options = {};
         if(this.option('selectionMode') === 'multiple') {
             if(!this._isPreserveSelectionMode) {
@@ -111,10 +194,108 @@ class FileManagerThumbnailListBox extends CollectionWidget {
         }
     }
 
+    // #region layoutModel
+
+    _resetLayoutModel() {
+        this._layoutModel = null;
+    }
+
+    _getLayoutModel() {
+        if(!this._layoutModel) {
+            this._layoutModel = this._createLayoutModel();
+        }
+        return this._layoutModel;
+    }
+
+    _createLayoutModel() {
+        if(this._getItemsLength() === 0) {
+            return null;
+        }
+
+        const item = this._getItems()[0];
+        const $item = this._getItemElementByItem(item);
+
+        const itemWidth = $item.outerWidth(true);
+        if(itemWidth === 0) {
+            return null;
+        }
+
+        const itemHeight = $item.outerHeight(true);
+
+        const viewPortWidth = this.$element().innerWidth();
+        const viewPortHeight = this._$scrollable.innerHeight();
+        const viewPortScrollTop = this._$scrollable.scrollTop();
+        const viewPortScrollBottom = viewPortScrollTop + viewPortHeight;
+
+        const itemPerRowCount = Math.floor(viewPortWidth / itemWidth);
+        const rowPerPageRate = viewPortHeight / itemHeight;
+
+        return {
+            itemWidth: itemWidth,
+            itemHeight: itemHeight,
+            viewPortWidth: viewPortWidth,
+            viewPortHeight: viewPortHeight,
+            viewPortScrollTop: viewPortScrollTop,
+            viewPortScrollBottom: viewPortScrollBottom,
+            itemPerRowCount: itemPerRowCount,
+            rowPerPageRate: rowPerPageRate
+        };
+    }
+
+    _createItemLayoutModel(index) {
+        const layout = this._getLayoutModel();
+        if(!layout) {
+            return null;
+        }
+
+        const itemRowIndex = Math.floor(index / layout.itemPerRowCount);
+        const itemColumnIndex = index % layout.itemPerRowCount;
+        const itemTop = itemRowIndex * layout.itemHeight;
+        const itemBottom = itemTop + layout.itemHeight;
+
+        return {
+            itemRowIndex: itemRowIndex,
+            itemColumnIndex: itemColumnIndex,
+            itemTop: itemTop,
+            itemBottom: itemBottom
+        };
+    }
+
+    _scrollToItem(item) {
+        const layout = this._getLayoutModel();
+        if(!layout) {
+            return;
+        }
+
+        const itemRowIndex = Math.floor(this._getIndexByItem(item) / layout.itemPerRowCount);
+        const itemTop = itemRowIndex * layout.itemHeight;
+        const itemBottom = itemTop + layout.itemHeight;
+
+        let newScrollTop = layout.viewPortScrollTop;
+
+        if(itemTop < layout.viewPortScrollTop) {
+            newScrollTop = itemTop;
+        } else if(itemBottom > layout.viewPortScrollBottom) {
+            newScrollTop = itemBottom - layout.viewPortHeight;
+        }
+
+        this._$scrollable.scrollTop(newScrollTop);
+    }
+
+    // #endregion layoutModel
+
     _focusOutHandler() {}
 
     getItems() {
         return this.option('items') || [];
+    }
+
+    _getItems() {
+        return this.option('items') || [];
+    }
+
+    _getItemsLength() {
+        return this._getItems().length;
     }
 
     getSelectedItems() {
@@ -125,6 +306,14 @@ class FileManagerThumbnailListBox extends CollectionWidget {
         return this._findItemElementByItem(item);
     }
 
+    _getItemElementByItem(item) {
+        return this._findItemElementByItem(item);
+    }
+
+    _getItemElementByIndex(index) {
+        return this.itemElements()[index];
+    }
+
     getItemByItemElement(itemElement) {
         return this.getItemByIndex(this.getIndexByItemElement(itemElement));
     }
@@ -133,20 +322,46 @@ class FileManagerThumbnailListBox extends CollectionWidget {
         return this._editStrategy.getNormalizedIndex(itemElement);
     }
 
-    getIndexByItem(item) {
-        return this._getIndexByItem(item);
-    }
+    // getIndexByItem(item) {
+    //     return this._getIndexByItem(item);
+    // }
+
+    // _getIndexByItem(item) {
+    //     return this._getIndexByItem(item);
+    // }
 
     getItemByIndex(index) {
-        return this.getItems()[index];
+        return this._getItems()[index];
+    }
+
+    _getItemByIndex(index) {
+        return this._getItems()[index];
     }
 
     getFocusedItem() {
         return this.getItemByItemElement(this.option('focusedElement'));
     }
 
+    _getFocusedItem() {
+        return this.getItemByItemElement(this.option('focusedElement'));
+    }
+
     setFocusedItem(item) {
         this.option('focusedElement', this.getItemElementByItem(item));
+    }
+
+    _focusItem(item, scrollToItem) {
+        this.setFocusedItem(item);
+        if(scrollToItem) {
+            this._scrollToItem(item);
+        }
+    }
+
+    _focusItemByIndex(index, scrollToItem, eventArgs) {
+        if(index >= 0 && index < this._getItemsLength()) {
+            const item = this._getItemByIndex(index);
+            this._focusItem(item, scrollToItem, eventArgs);
+        }
     }
 
     selectAll() {
@@ -164,6 +379,10 @@ class FileManagerThumbnailListBox extends CollectionWidget {
 
     clearFocus() {
         this.option('focusedElement', null);
+    }
+
+    getItemClass() {
+        return this._itemClass();
     }
 
 }
